Front end for p4merge for use from within Visual Studio.
When calling an external merge tool, Visual Studio supplies
an argument for where the merged file should be saved, but
it does not create this file. Upon startup, p4merge checks
for the existence of such a file, and if one does not exist,
it returns an error.

This program takes the parameter passed from Visual Studio,
creates an empty file, and in turn calls p4merge.

Note: when configuring this from within Visual Studio,
the correct order of the arguments is %3 %1 %2 %4.